requirements.txt
chromadb==0.6.3
matplotlib==3.10.0
Pillow==11.1.0
sentence_transformers==3.4.1
torch==2.6.0
transformers==4.46.1
unstructured==0.16.20
